Comentado soporte para Ktor hasta que funcione 100%: implementar fábricas de clientes HTTP para Android, escritorio y WebAssembly, y actualizar las dependencias en build.gradle.kts para la serialización y negociación de contenido.
