+++ composeApp/build.gradle.kts
@@ -42,7 +42,28 @@ kotlin {
         }
         binaries.executable()
     }
-    
+
+    /*
+    js(IR) {
+        moduleName = "composeApp"
+        browser {
+            val rootDirPath = project.rootDir.path
+            val projectDirPath = project.projectDir.path
+            commonWebpackConfig {
+                outputFileName = "composeApp.js"
+                devServer = (devServer ?: KotlinWebpackConfig.DevServer()).apply {
+                    static = (static ?: mutableListOf()).apply {
+                        // Serve sources to debug inside browser
+                        add(rootDirPath)
+                        add(projectDirPath)
+                    }
+                }
+            }
+        }
+        binaries.executable()
+    }
+     */
+
     sourceSets {
         val desktopMain by getting
         // val wasmJsMain by getting
@@ -56,7 +77,7 @@ kotlin {
             implementation(libs.kotlinx.serialization)
 
             // Ktor
-            implementation(libs.ktor.client.okhttp)
+            //implementation(libs.ktor.client.okhttp)
         }
         commonMain.dependencies {
             implementation(compose.runtime)
@@ -75,12 +96,12 @@ kotlin {
             implementation(libs.kotlinx.serialization)
 
             // Ktor
-            implementation(libs.ktor.client.core)
-            implementation(libs.ktor.client.content.negotiation)
-            implementation(libs.ktor.serialization.kotlinx.json)
+            //implementation(libs.ktor.client.core)
+            //implementation(libs.ktor.client.content.negotiation)
+            //implementation(libs.ktor.serialization.kotlinx.json)
 
             // Ktor for JS
-            implementation(libs.ktor.client.js)
+            //implementation(libs.ktor.client.js)
         }
         desktopMain.dependencies {
             implementation(compose.desktop.currentOs)
@@ -90,20 +111,39 @@ kotlin {
             implementation(libs.decompose.jetbrains)
 
             // Ktor
-            implementation(libs.ktor.client.cio.jvm)
+            //implementation(libs.ktor.client.cio.jvm)
+            //implementation(libs.ktor.client.cio)
         }
+        /*
+        val jvmMain by getting {
+            dependencies {
+                implementation(libs.ktor.client.cio) // JVM-specific Ktor client
+                // Other JVM-specific dependencies
+            }
+        }
+         */
         val wasmJsMain by getting {
             dependencies {
                 //implementation("org.jetbrains.compose.web:compose-web-router:1.0.0-beta6")
                 implementation(libs.kotlinx.serialization)
 
                 // Ktor
-                implementation(libs.ktor.client.js)
-                implementation(libs.ktor.client.content.negotiation) // Faltaba
-                implementation(libs.ktor.serialization.kotlinx.json) // Faltaba
-                implementation(libs.kotlinx.serialization) // Requerido
+                //implementation(libs.ktor.client.js)
+                //implementation(libs.ktor.client.content.negotiation.v238) // Faltaba
+                //implementation(libs.ktor.client.core.v238) // ¡Versión 2.3.8+!
+                //implementation(libs.ktor.serialization.kotlinx.json) // Faltaba
+                //implementation(libs.ktor.serialization.kotlinx.json.v238) // Requerido
             }
         }
+        /*
+        val jsMain by getting {
+            dependencies {
+                implementation(libs.ktor.client.js) // Usa el cliente JS
+                implementation(libs.ktor.client.content.negotiation.v238)
+                implementation(libs.ktor.serialization.kotlinx.json.v238)
+            }
+        }
+         */
     }
 }
