+++ composeApp/src/androidMain/kotlin/org/connexuss/project/androidHttpClientFactory.kt
@@ -1,0 +1,16 @@
+package org.connexuss.project
+/*
+import io.ktor.client.HttpClient
+import io.ktor.client.engine.okhttp.OkHttp
+import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
+import io.ktor.serialization.kotlinx.json.json
+import kotlinx.serialization.json.Json
+
+actual fun createHttpClient(): HttpClient = HttpClient(OkHttp) {
+    install(ContentNegotiation) {
+        json(Json {
+            ignoreUnknownKeys = true
+        })
+    }
+}
+ */
