+++ composeApp/src/commonMain/kotlin/org/connexuss/project/App.kt
@@ -9,9 +9,6 @@ import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import org.connexuss.project.navegacion.Navegacion
 import org.jetbrains.compose.ui.tooling.preview.Preview
-import io.ktor.client.*
-import io.ktor.client.request.*
-import io.ktor.client.statement.*
 
 @Composable
 @Preview
@@ -34,17 +31,3 @@ fun App() {
         )
     }
 }
-/*
-// Declaración "expect" para crear el cliente HTTP en cada plataforma
-expect fun createHttpClient(): HttpClient
-
-// Función compartida para hacer una petición GET a una URL (por ejemplo, a la API REST de Firebase)
-suspend fun fetchFirebaseData(url: String): String {
-    val client = createHttpClient()
-    // Realiza una solicitud GET y obtiene el cuerpo como texto
-    val response: HttpResponse = client.get(url)
-    val text = response.bodyAsText()
-    client.close()
-    return text
-}
- */
