+++ composeApp/src/desktopMain/kotlin/org/connexuss/project/desktopHttpClientFactory.kt
@@ -1,0 +1,16 @@
+package org.connexuss.project
+/*
+import io.ktor.client.*
+import io.ktor.client.engine.cio.*
+import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
+import io.ktor.serialization.kotlinx.json.json
+import kotlinx.serialization.json.Json
+
+actual fun createHttpClient(): HttpClient = HttpClient(CIO) {
+    install(ContentNegotiation) {
+        json(Json {
+            ignoreUnknownKeys = true
+        })
+    }
+}
+ */
