+++ composeApp/src/desktopMain/kotlin/org/connexuss/project/main.kt
@@ -2,11 +2,6 @@ package org.connexuss.project
 
 import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.application
-import io.ktor.client.*
-import io.ktor.client.engine.cio.*
-import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
-import io.ktor.serialization.kotlinx.json.json
-import kotlinx.serialization.json.Json
 
 fun main() = application {
     Window(
@@ -17,10 +12,3 @@ fun main() = application {
     }
 }
 
-actual fun createHttpClient(): HttpClient = HttpClient(CIO) {
-    install(ContentNegotiation) {
-        json(Json {
-            ignoreUnknownKeys = true
-        })
-    }
-}
