+++ gradle.properties
@@ -1,6 +1,7 @@
 #Kotlin
 kotlin.code.style=official
 kotlin.daemon.jvmargs=-Xmx2048M
+kotlin.wasm.enabled=true
 
 #Gradle
 org.gradle.jvmargs=-Xmx2048M -Dfile.encoding=UTF-8
@@ -10,3 +11,4 @@ org.gradle.jvmargs=-Xmx2048M -Dfile.encoding=UTF-8
 android.nonTransitiveRClass=true
 android.useAndroidX=true
 org.gradle.configuration-cache=true
+org.jetbrains.compose.experimental.jscanvas.enabled=true
