+++ gradle/libs.versions.toml
@@ -24,6 +24,9 @@ kotlinxCrypto = "0.1.0"
 firebaseFirestoreKtx = "25.1.2"
 kotlinxDatetime = "0.6.0"
 decompose = "2.2.0-alpha03"
+ktorClientContentNegotiation = "2.3.8"
+ktorClientCore = "2.3.8"
+ktorSerializationKotlinxJson = "2.3.8"
 navigation-compose = "2.8.0-alpha10"
 extensionsComposeJetbrains = "2.2.0-alpha03"
 kotlinxSerialization = "1.7.0"
@@ -31,7 +34,7 @@ foundationAndroid = "1.7.8"
 animationCoreLint = "1.8.0-beta03"
 
 # Ktor
-ktor = "2.3.7"
+ktor = "2.3.8"
 
 [libraries]
 arkivanov-decompose-v080 = { module = "com.arkivanov.decompose:decompose", version.ref = "arkivanovDecompose" }
@@ -55,6 +58,9 @@ firebase-firestore-ktx = { group = "com.google.firebase", name = "firebase-fires
 kotlinx-datetime = { module = "org.jetbrains.kotlinx:kotlinx-datetime", version.ref = "kotlinxDatetime" }
 decompose = { module = "com.arkivanov.decompose:decompose", version.ref = "decompose" }
 decompose-compose = { module = "com.arkivanov.decompose:extensions-compose-jetpack", version.ref = "decompose" }
+ktor-client-content-negotiation-v238 = { module = "io.ktor:ktor-client-content-negotiation", version.ref = "ktorClientContentNegotiation" }
+ktor-client-core-v238 = { module = "io.ktor:ktor-client-core", version.ref = "ktorClientCore" }
+ktor-serialization-kotlinx-json-v238 = { module = "io.ktor:ktor-serialization-kotlinx-json", version.ref = "ktorSerializationKotlinxJson" }
 navigation-compose = { module = "org.jetbrains.androidx.navigation:navigation-compose", version.ref = "navigation-compose" }
 #decompose-jetbrains = { module = "com.arkivanov.decompose:extensions-jetbrains-compose", version.ref = "extensionsComposeJetbrains" }
 decompose-jetbrains = { module = "com.arkivanov.decompose:extensions-compose-jetbrains", version.ref = "extensionsComposeJetbrains" }
@@ -67,7 +73,8 @@ ktor-client-core = { module = "io.ktor:ktor-client-core", version.ref = "ktor" }
 ktor-client-content-negotiation = { module = "io.ktor:ktor-client-content-negotiation", version.ref = "ktor" }
 ktor-serialization-kotlinx-json = { module = "io.ktor:ktor-serialization-kotlinx-json", version.ref = "ktor" }
 ktor-client-okhttp = { module = "io.ktor:ktor-client-okhttp", version.ref = "ktor" }
-ktor-client-cio-jvm = { module = "io.ktor:ktor-client-cio-jvm", version.ref = "ktor" }
+#ktor-client-cio-jvm = { module = "io.ktor:ktor-client-cio-jvm", version.ref = "ktor" }
+ktor-client-cio = { module = "io.ktor:ktor-client-cio", version.ref = "ktor" }
 ktor-client-js = { module = "io.ktor:ktor-client-js", version.ref = "ktor" }
 
 [plugins]
