+++ settings.gradle.kts
@@ -29,6 +29,7 @@ dependencyResolutionManagement {
         mavenCentral()
         maven { url = uri("https://maven.pkg.jetbrains.space/public/p/compose/dev") }
         maven{ url = uri("https://maven.pkg.jetbrains.space/kotlin/p/wasm/experimental") }
+        maven("https://repo.repsy.io/mvn/arkivanov/public") // Repositorio de Arkivanov
     }
 }
